Move rubric relaetd types into a separate file
