+++ pxtblocks/code-validation/rubricCriteria.ts
@@ -1,34 +1,4 @@
 namespace pxt.blocks {
-
-    // todo thsparks : move these types elsewhere...
-    export type CriteriaParameterPicker = "blocksPicker" | "numberPicker";
-
-    export interface CriteriaParameter {
-        name: string;
-        type: string;
-        picker: CriteriaParameterPicker;
-        path: string;
-    }
-    export interface CriteriaParameterValue {
-        name: string;
-        value: any;
-    }
-
-    export interface CatalogCriteria {
-        id: string;
-        use: string;
-        template: string;
-        description: string | undefined;
-        docPath: string | undefined;
-        parameters: CriteriaParameter[] | undefined;
-    }
-
-    export interface CriteriaInstance {
-        catalogCriteriaId: string;
-        instanceId: string;
-        params: CriteriaParameterValue[] | undefined;
-    }
-
     interface BlockSet {
         blocks: string[];
         count: number;
